requirements.txt
django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-decouple>=3.8
requests>=2.31
pillow>=10.0
gunicorn>=21.0
whitenoise>=6.6
